test: Consolidate bug-fix tests into main suites and add multi-table e2e coverage
Merge isolated *_bug_fixes_test.cpp files into their canonical test suites
and expand e2e coverage with multi-table isolation and search freshness tests.

C++ test consolidation:
- Delete tests/index/index_bug_fixes_test.cpp and distribute its tests:
  - Bug #14/#15 (empty posting list cleanup after RemoveDocument/UpdateDocument)
    and EstimatePostingSize basic test moved into index_basic_test.cpp
  - Bug #16 (SearchAnd large posting list materialization) tests moved into
    index_search_test.cpp under the IndexTest suite
  - EstimatePostingSize concurrent access test moved into index_concurrent_test.cpp
  - CMakeLists.txt updated to remove the index_bug_fixes_test target
- Delete tests/loader/initial_loader_bug_fixes_test.cpp; its batch processing
  tests (Bug #4, Bug #5, size assertions) moved into initial_loader_query_test.cpp
  with the mygramdb_storage and mygramdb_index link dependencies transferred there
- Delete tests/mysql/rows_parser_bug_fixes_test.cpp; its content merged into
  rows_parser_test.cpp
- Rename tests/server/search_handler_bug_fixes_test.cpp to
  tests/server/search_handler_test.cpp as the primary search handler test file

New e2e tests:
- e2e/tests/multi_table/test_multi_table_isolation.py (451 lines, pytestmark:
  multi_table): 9 test cases verifying that operations on one table do not
  affect the other table's index
  - TestMultiTableSearch: both_tables_searchable, no_cross_contamination_search,
    filter_isolation
  - TestMultiTableReplication: insert_both_tables_independently,
    update_one_table_no_impact_on_other, delete_one_table_no_impact_on_other
  - TestMultiTableDDL: drop_one_table_other_still_works,
    truncate_one_table_other_intact, drop_and_recreate_table_search_recovers,
    sync_one_table_no_impact_on_other
- e2e/tests/replication/test_search_freshness.py (539 lines, pytestmark:
  replication): 13 test cases verifying SEARCH (not COUNT) reflects replication
  events, structured to isolate index-update bugs from cache-invalidation bugs
  - TestSearchFreshnessNocache: insert_appears_in_search,
    update_text_changes_search_results, delete_removes_from_search,
    bulk_insert_all_searchable
  - TestSearchFreshnessWithCache: insert_appears_in_search_with_cache,
    update_invalidates_cached_search, delete_invalidates_cached_search,
    filter_update_reflected_in_cached_search
  - TestSearchFreshnessCacheToggle: disable_during_replication_then_reenable,
    insert_while_cache_off_then_reenable, delete_while_cache_off_then_reenable,
    rapid_toggle_during_writes

e2e infrastructure:
- Add products table to e2e/docker/mygramdb-test.yaml for multi-table tests
  (primary_key: id, text_source: description, required_filters: enabled=1,
  filters: status, category)
- Update e2e/conftest.py seed_data fixture to seed both articles (100 rows)
  and products (50 rows), updating the doc_count threshold from 90 to 130
- Fix e2e/lib/mygramdb_client.py FILTER syntax in search() and count() from
  "FILTER key value" to "FILTER key=value" to match the actual protocol format
