+++ e2e/docker/mygramdb-test.yaml
@@ -31,6 +31,28 @@ tables:
     ngram_size: 2
     kanji_ngram_size: 1
 
+  - name: "products"
+    primary_key: "id"
+    text_source:
+      column: "description"
+    required_filters:
+      - name: "enabled"
+        type: "int"
+        op: "="
+        value: 1
+        bitmap_index: false
+    filters:
+      - name: "status"
+        type: "int"
+        dict_compress: false
+        bitmap_index: false
+      - name: "category"
+        type: "string"
+        dict_compress: false
+        bitmap_index: false
+    ngram_size: 2
+    kanji_ngram_size: 1
+
 build:
   batch_size: 5000
   parallelism: 2
