+++ tests/index/CMakeLists.txt
@@ -170,21 +170,6 @@ else()
   )
 endif()
 
-# Index bug fix tests (TDD)
-add_executable(index_bug_fixes_test
-  index_bug_fixes_test.cpp
-)
-
-target_link_libraries(index_bug_fixes_test
-  mygramdb_index
-  GTest::gtest_main
-)
-
-gtest_discover_tests(index_bug_fixes_test
-  DISCOVERY_MODE POST_BUILD
-  DISCOVERY_TIMEOUT 30
-)
-
 # Empty document return value tests
 add_executable(index_empty_doc_test
   index_empty_doc_test.cpp
