+++ tests/server/CMakeLists.txt
@@ -540,18 +540,18 @@ gtest_discover_tests(snapshot_scheduler_test
   PROPERTIES RESOURCE_LOCK storage_io
 )
 
-# SearchHandler bug fix tests (TDD)
+# SearchHandler tests
 
-add_executable(search_handler_bug_fixes_test
-  search_handler_bug_fixes_test.cpp
+add_executable(search_handler_test
+  search_handler_test.cpp
 )
 
-target_link_libraries(search_handler_bug_fixes_test
+target_link_libraries(search_handler_test
   mygramdb_server
   GTest::gtest_main
 )
 
-gtest_discover_tests(search_handler_bug_fixes_test
+gtest_discover_tests(search_handler_test
   DISCOVERY_MODE POST_BUILD
   DISCOVERY_TIMEOUT 30
   PROPERTIES RESOURCE_LOCK server_state
